# Stockwave-Postman-Collection-Json2OpenApi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Stockwave",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "variables": {}
    }
  ],
  "paths": {
...
    "/professor": {
      "get": {
        "tags": [
          "Professor"
        ],
        "summary": "Listar Professores",
        "operationId": "ListarProfessores",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "Professor"
        ],
        "summary": "Cadastrar Professor",
        "operationId": "CadastrarProfessor",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CadastrarProfessorRequest"
                  },
                  {
                    "example": {
                      "cpf_usuario": "46145226955",
                      "email_usuario": "TESTANDO",
                      "nome_usuario": "TESTANDO"
                    }
                  }
                ]
              },
              "example": {
                "cpf_usuario": "46145226955",
                "email_usuario": "TESTANDO",
                "nome_usuario": "TESTANDO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/professor/8": {
      "put": {
        "tags": [
          "Professor"
        ],
        "summary": "Atualizar Professor",
        "operationId": "AtualizarProfessor",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AtualizarProfessorRequest"
                  },
                  {
                    "example": {
                      "cpf_usuario": "12345678910",
                      "email_usuario": "PROF@GMAIL.COM",
                      "id_usuario": 8,
                      "nome_usuario": "PROFESSOR TESTE"
                    }
                  }
                ]
              },
              "example": {
                "cpf_usuario": "12345678910",
                "email_usuario": "PROF@GMAIL.COM",
                "id_usuario": 8,
                "nome_usuario": "PROFESSOR TESTE"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "delete": {
        "tags": [
          "Professor"
        ],
        "summary": "Deletar Professor",
        "operationId": "DeletarProfessor",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/resposta": {
      "get": {
        "tags": [
          "Resposta"
        ],
        "summary": "Listar Respostas",
        "operationId": "ListarRespostas",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "Resposta"
        ],
        "summary": "Cadastrar Resposta",
        "operationId": "CadastrarResposta",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CadastrarRespostaRequest"
                  },
                  {
                    "example": {
                      "resposta": "H"
                    }
                  }
                ]
              },
              "example": {
                "resposta": "H"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/resposta/F": {
      "put": {
        "tags": [
          "Resposta"
        ],
        "summary": "Atualizar Resposta",
        "operationId": "AtualizarResposta",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/AtualizarRespostaRequest"
                  },
                  {
                    "example": {
                      "resposta": "A"
                    }
                  }
                ]
              },
              "example": {
                "resposta": "A"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AtualizarProfessorRequest": {
        "title": "AtualizarProfessorRequest",
        "required": [
          "cpf_usuario",
          "email_usuario",
          "id_usuario",
          "nome_usuario"
        ],
        "type": "object",
        "properties": {
          "cpf_usuario": {
            "type": "string"
          },
          "email_usuario": {
            "type": "string"
          },
          "id_usuario": {
            "type": "integer",
            "format": "int32"
          },
          "nome_usuario": {
            "type": "string"
          }
        },
        "example": {
          "cpf_usuario": "12345678910",
          "email_usuario": "PROF@GMAIL.COM",
          "id_usuario": 8,
          "nome_usuario": "PROFESSOR TESTE"
        }
      },
      "CadastrarProfessorRequest": {
        "title": "CadastrarProfessorRequest",
        "required": [
          "cpf_usuario",
          "email_usuario",
          "nome_usuario"
        ],
        "type": "object",
        "properties": {
          "cpf_usuario": {
            "type": "string"
          },
          "email_usuario": {
            "type": "string"
          },
          "nome_usuario": {
            "type": "string"
          }
        },
        "example": {
          "cpf_usuario": "46145226955",
          "email_usuario": "TESTANDO",
          "nome_usuario": "TESTANDO"
        }
      },
...
      "AtualizarRespostaRequest": {
        "title": "AtualizarRespostaRequest",
        "required": [
          "resposta"
        ],
        "type": "object",
        "properties": {
          "resposta": {
            "type": "string"
          }
        },
        "example": {
          "resposta": "A"
        }
      },
      "CadastrarRespostaRequest": {
        "title": "CadastrarRespostaRequest",
        "required": [
          "resposta"
        ],
        "type": "object",
        "properties": {
          "resposta": {
            "type": "string"
          }
        },
        "example": {
          "resposta": "H"
        }
      },
...
    }
  },
...
}